generator/requirements.txt
numpy
opencv-contrib-python
pillow
torch
torchvision

